-Agregación de rutas para admisnitradores -Agregación de componentes CRUD administrador,usario, registro, adopcion, mascotas -Modificación de crud administrador
